Day1 Web2/app/requirement.txt
backports-abc==0.5
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cssselect==1.0.3
futures==3.2.0
idna==2.6
lxml
pycparser==2.18
pyquery==1.4.0
requests==2.18.4
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.7
tornado==5.0.2
urllib3==1.22
PyJWT==1.6.1

